Update README.md
The equality relational operator

Step 1

Run the program several times using a different input each time. Does the program do what you think it was meant to be doing? If not, locate the error and fix it.

Step 2

Modify the program so that the user inputs both values to be tested for equality. Make sure you have a prompt for each input. Test the program with pairs of values that are the same and that are different.

Step 3

Modify the program so that when the numbers are the same it prints the following two lines:

The values are the same.
Hey that’s a coincidence!

Step 4

Modify the program by replacing the two if statements with a single if/else statement. Run the program again to test the results.